renaming and sorting naming conventions
